requirements.txt
configparser
flake8
mock
prometheus_client
python-neutronclient
python-novaclient
python-cinderclient
requests
PyYAML